@@ -10,32 +10,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.9",
-    "@angular/cdk": "^14.2.7",
-    "@angular/common": "^15.2.9",
-    "@angular/compiler": "^15.2.9",
-    "@angular/core": "^15.2.9",
-    "@angular/forms": "^15.2.9",
-    "@angular/platform-browser": "^15.2.9",
-    "@angular/platform-browser-dynamic": "^15.2.9",
-    "@angular/router": "^15.2.9",
-    "@nebular/auth": "^10.0.0",
-    "@nebular/eva-icons": "10.0.0",
-    "@nebular/security": "^10.0.0",
-    "@nebular/theme": "^10.0.0",
+    "@angular/animations": "^16.0.2",
+    "@angular/cdk": "^16.0.1",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.2",
+    "@angular/core": "^16.0.2",
+    "@angular/forms": "^16.0.2",
+    "@angular/platform-browser": "^16.0.2",
+    "@angular/platform-browser-dynamic": "^16.0.2",
+    "@angular/router": "^16.0.2",
+    "@nebular/auth": "^11.0.0",
+    "@nebular/eva-icons": "^11.0.0",
+    "@nebular/security": "^11.0.0",
+    "@nebular/theme": "^11.0.0",
     "@syncfusion/ej2-angular-schedule": "^21.2.5",
     "bootstrap": "^5.2.3",
     "eva-icons": "^1.1.2",
     "primeng": "^16.0.0-rc.2",
-    "rxjs": "~7.5.0",
+    "rxjs": "^7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.8",
-    "@angular/cli": "~15.2.8",
-    "@angular/compiler-cli": "^15.2.9",
-    "@schematics/angular": "~14.1.3",
+    "@angular-devkit/build-angular": "^16.0.2",
+    "@angular/cli": "~16.0.2",
+    "@angular/compiler-cli": "^16.0.2",
+    "@schematics/angular": "~16.0.2",
     "@types/jasmine": "~4.0.0",
     "jasmine-core": "~4.2.0",
     "karma": "~6.4.0",
