@@ -1,10 +1,10 @@
 import {Injectable} from '@angular/core'
-import {CanActivate,Router} from '@angular/router'
+import { Router } from '@angular/router'
 import { NbAuthService } from '@nebular/auth';
 import { tap } from 'rxjs';
 
 @Injectable()
-export class AuthGuard implements CanActivate {
+export class AuthGuard  {
 
     constructor(
         private authService: NbAuthService,
